Image annotation and MongoDB updates.
The UI was updated so that the document labels were on the same tab
as the images to make image annotation easier.  The eve container
was also updated to install MongoDB from a .tar.gz installer instead
of a package.  This is because the packages now require systemctl
to be installed and it is not present in a Ubuntu container.